+++ include/wx/qt/clrpicker.h
@@ -28,7 +28,7 @@ public:
                    const wxColour& initial = *wxBLACK,
                    const wxPoint& pos = wxDefaultPosition,
                    const wxSize& size = wxDefaultSize,
-                   long style = wxCLRBTN_DEFAULT_STYLE,
+                   unsigned int style = wxCLRBTN_DEFAULT_STYLE,
                    const wxValidator& validator = wxDefaultValidator,
                    const wxString& name = wxASCII_STR(wxColourPickerWidgetNameStr));
 
@@ -37,7 +37,7 @@ public:
                 const wxColour& initial = *wxBLACK,
                 const wxPoint& pos = wxDefaultPosition,
                 const wxSize& size = wxDefaultSize,
-                long style = wxCLRBTN_DEFAULT_STYLE,
+                unsigned int style = wxCLRBTN_DEFAULT_STYLE,
                 const wxValidator& validator = wxDefaultValidator,
                 const wxString& name = wxASCII_STR(wxColourPickerWidgetNameStr));
